+++ src/server/server.js
@@ -104,6 +104,7 @@ function httpRoute(req, res) {
                 window.IP=${JSON.stringify(ip)}
                 window.ENV=${JSON.stringify(clientOptions)}
                 window.READ_ONLY=${JSON.stringify(settings.read('read-only'))}
+                window.KIOSK=${settings.read('read-only') && settings.read('fullscreen') ? 'true' : 'false'}
                 window.AUTH=\`${(req.headers.authorization || '').replace('Basic ', '').replace('/', '_')}\`
             </script></body>`))
           .pipe(res)