+++ src/client/widgets/sliders/range.js
@@ -384,6 +384,11 @@ class Range extends Fader {
 
         }
 
+        if (this.shouldDrawPips) {
+            this.drawPips()
+            this.shouldDrawPips = false
+        }
+
         if (this.getProp('pips')) {
             this.ctx.globalAlpha = 1
             this.ctx.drawImage(this.pips, 0, 0)